feat(search): integrate endianness selection and enhance search UI
